refactor(team): split broker_office_members out of broker_office_channels.go
Phase 4. The 392-LOC handleOfficeMembers handler had a TODO at the call
site asking for parser/applier separation; this PR delivers it.

Split:
  - HTTP entry: handleOfficeMembers (method routing only)
  - GET path: serveOfficeMemberList (with officeMemberListEntry type)
  - POST path: serveOfficeMemberMutation (parses + dispatches by action)
  - Action bodies, each holding b.mu and reviewable in isolation:
      - createOfficeMemberLocked: openclaw bridge attach BEFORE persist;
        seeds new hire into every non-DM channel; clears stale Disabled
      - updateOfficeMemberLocked: provider switch reconciles bridge
        (attach-before-detach) so attach failures preserve old binding
      - removeOfficeMemberLocked: best-effort openclaw detach; cleans
        Members + Disabled across non-DM channels; releases owned tasks

  - internal/team/broker_office_channels.go: -393 LOC (1657 -> 1264)
  - internal/team/broker_office_members.go: +456 LOC (new)

The original handler's behavior is preserved exactly. Each action body
is now a candidate for a dedicated unit test (today they were only
covered through HTTP-level tests).
